% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_tests.R
\name{cd}
\alias{cd}
\title{Title}
\usage{
cd(data, var, id = NULL, time = NULL, na.action = na.omit, ...)
}
\arguments{
\item{data}{dataset}

\item{var}{variable to be tested}

\item{id}{panel data id indicator}

\item{time}{panel data time indicator}

\item{na.action}{na actions}

\item{...}{other argunents}
}
\value{
CD test statistic and p-value
}
\description{
Title
}
