% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce_mean_group.R
\name{coef.cce_mean_group}
\alias{coef.cce_mean_group}
\title{Coefficients method for CCE-MG estimator}
\usage{
\method{coef}{cce_mean_group}(object)
}
\arguments{
\item{object}{An object of class 'cce_mean_group'.}
}
\description{
Coefficients method for CCE-MG estimator
}
