% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{compute_vcov}
\alias{compute_vcov}
\title{Compute Variance-Covariance Matrix}
\usage{
compute_vcov(coef_matrix, method = "standard")
}
\arguments{
\item{coef_matrix}{A numeric matrix where each row corresponds to individual estimates of parameters.}

\item{method}{A character string indicating the method to compute the covariance matrix. Options are "standard" (default) or "robust".}
}
\value{
A variance-covariance matrix.
}
\description{
Computes the variance-covariance matrix for a given matrix of estimated coefficients.
}
