% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm-package.R
\docType{package}
\name{csdm-package}
\alias{csdm-package}
\title{The 'csdm' package.}
\description{
A tool for catch saturation-depletion models with R and STAN
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
