% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_extract.R
\name{csdm_extract}
\alias{csdm_extract}
\alias{nobs.csdm_fit}
\alias{model.frame.csdm_fit}
\alias{model.matrix.csdm_fit}
\alias{terms.csdm_fit}
\alias{formula.csdm_fit}
\alias{df.residual.csdm_fit}
\title{Extract fitted model data and sample information}
\usage{
\method{nobs}{csdm_fit}(object, ...)

\method{model.frame}{csdm_fit}(formula, ...)

\method{model.matrix}{csdm_fit}(object, ...)

\method{terms}{csdm_fit}(x, ...)

\method{formula}{csdm_fit}(x, ...)

\method{df.residual}{csdm_fit}(object, ...)
}
\arguments{
\item{object, x}{A csdm_fit object.}

\item{...}{Further arguments.}
}
\description{
The model frame and economic design matrix contain only estimated observations,
in panel order. formula() and terms() include constructed economic lags.
nobs() counts estimated observations. df.residual() returns Inf for asymptotic
normal MG inference; individual regression degrees of freedom are in object$units.
}
