% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_stan.R
\name{csdm_stan}
\alias{csdm_stan}
\title{Bayesian catch saturation-depletion model (stan routine)}
\usage{
csdm_stan(data, ...)
}
\arguments{
\item{data}{\strong{[list]} Data to run the model, passed as a list.}

\item{...}{Additional arguments to pass to \link[rstan:stanmodel-method-sampling]{sampling}.}
}
\value{
An object of class \code{stanfit} returned by \link[rstan:stanmodel-method-sampling]{sampling}.
}
\description{
Stan routine for the catch saturation-depletion model.
}
