% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_tidy.R
\name{csdm_tidy}
\alias{csdm_tidy}
\alias{tidy.csdm_fit}
\alias{glance.csdm_fit}
\alias{augment.csdm_fit}
\title{Tidy model coefficients, fit statistics, and observations}
\usage{
\method{tidy}{csdm_fit}(
  x,
  component = c("all", "levels", "adjustment", "long_run"),
  conf.int = FALSE,
  conf.level = 0.95,
  ...
)

\method{glance}{csdm_fit}(x, ...)

\method{augment}{csdm_fit}(x, data = x$data, newdata = NULL, ...)
}
\arguments{
\item{x}{A csdm_fit object.}

\item{component}{Parameter component, as in coef().}

\item{conf.int}{Include normal-approximation confidence intervals.}

\item{conf.level}{Confidence level between zero and one.}

\item{...}{Further arguments; currently unused.}

\item{data}{Original fitting data, in its original order.}

\item{newdata}{Not supported.}
}
\description{
Tidy model coefficients, fit statistics, and observations
}
