% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_specs.R
\name{csdm_vcov}
\alias{csdm_vcov}
\title{Specification: Mean-group variance-covariance estimator}
\usage{
csdm_vcov(type = "mg", ...)
}
\arguments{
\item{type}{Must be \code{"mg"}, the implemented mean-group variance estimator.}

\item{...}{Must be empty; additional variance estimators are not implemented.}
}
\value{
A spec object (list) used by csdm().
}
\description{
Specification: Mean-group variance-covariance estimator
}
