% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Sample dataset for csdm}
\format{
A data frame.
}
\usage{
data
}
\description{
A small panel dataset shipped with the package (loaded from
`data/xtdcce2_sample_dataset.rda`).
}
\keyword{datasets}
