% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delury_lobster}
\alias{delury_lobster}
\title{Delury lobster data}
\format{
A data frame with 33 rows and 7 variables:
\describe{
\item{date}{Day number.}
\item{catch_pounds}{Lobster catch, in pounds.}
\item{traps}{Number of traps.}
\item{cpue}{Catch per unit effort, such as \code{cpue = catch_pounds/traps}}
\item{traps_cum}{Number of traps, cumulated.}
\item{cpue_l}{}
\item{cpue_l10}{}
}
}
\source{
\url{https://www.jstor.org/stable/3001390}
}
\usage{
delury_lobster
}
\description{
A dataset containing catch data for lobster from DeLury, D. B. (1947). On the
estimation of biological populations. Biometrics, 3, 145–167.
}
\keyword{datasets}
