% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_internal_helpers.R
\name{.csdm_significance_stars}
\alias{.csdm_significance_stars}
\title{R-style significance codes for p-values (see ?summary.lm)}
\usage{
.csdm_significance_stars(p)
}
\description{
R-style significance codes for p-values (see ?summary.lm)
}
