% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_biomass}
\alias{get_biomass}
\title{Extract biomass from a csdm_fit object}
\usage{
get_biomass(csdm_fit)
}
\arguments{
\item{csdm_fit}{\strong{[csdm_fit]} The csdm_fit object.}
}
\value{
A data.frame with the model biomass
}
\description{
Extract biomass from a csdm_fit object
}
\examples{
\dontrun{
get_biomass(model_fit)
}

}
