% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_cpue_fit}
\alias{get_cpue_fit}
\title{Extract fit from a csdm_fit object}
\usage{
get_cpue_fit(csdm_fit)
}
\arguments{
\item{csdm_fit}{\strong{[csdm_fit]} The csdm_fit object.}
}
\value{
A data.frame with the model fit.
}
\description{
Extract fit from a csdm_fit object
}
\examples{
\dontrun{
get_cpue_fit(model_fit)
}

}
