% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_fit_class.R
\name{new_csdm_fit}
\alias{new_csdm_fit}
\title{Create a csdm_fit object}
\usage{
new_csdm_fit(data, fit)
}
\arguments{
\item{data}{\strong{[data.frame]} The model data.}

\item{fit}{\strong{[stanfit]} The model fit.}
}
\value{
An object of S3 class csdm_fit.
}
\description{
Creates a csdm_fit object.
}
\examples{
\dontrun{
csdm_fit <- new_csdm_fit(data = data, fit = model_fit)
}

}
