% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.csdm_fit}
\alias{plot.csdm_fit}
\title{Plot model output}
\usage{
\method{plot}{csdm_fit}(x, y, ...)
}
\arguments{
\item{x}{\strong{[csdm_fit]} The csdm_fit object.}

\item{y}{Not used}

\item{...}{Not used}
}
\value{
A ggplot.
}
\description{
Plots the output of the model.
}
\examples{
\dontrun{
plot(delury_model)
plot_biomass(delury_model)
}

}
