% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Make a biomass plot}
\usage{
plot_biomass(x)
}
\arguments{
\item{x}{\strong{[csdm_fit]} The csdm_fit object.}
}
\value{
A ggplot of biomass.
}
\description{
Produces a generic plot of biomass from the model
}
\examples{
\dontrun{
plot(delury_model)
plot_biomass(delury_model)
}

}
