% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{predict.cce_mean_group}
\alias{predict.cce_mean_group}
\title{Predict cce_mg}
\usage{
\method{predict}{cce_mean_group}(object, data = NULL, id = NULL, time = NULL)
}
\arguments{
\item{object}{an cee-mg model}

\item{data}{optional dataset}

\item{id}{id panel data id indicator}

\item{time}{time panel data time indicator}
}
\value{
predictions
}
\description{
Predict cce_mg
}
