% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{predict_y}
\alias{predict_y}
\title{Predict Y using estimated coefficients}
\usage{
predict_y(coef_vector, X_matrix)
}
\arguments{
\item{coef_vector}{A numeric vector of estimated coefficients.}

\item{X_matrix}{A numeric matrix of independent variables (including intercept if applicable).}
}
\value{
A numeric vector of predicted values.
}
\description{
Predicts the dependent variable using estimated coefficients and input features.
}
