% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce_mean_group.R
\name{print.cce_mean_group}
\alias{print.cce_mean_group}
\title{Print method for CCE-MG estimator}
\usage{
\method{print}{cce_mean_group}(x)
}
\arguments{
\item{x}{An object of class 'cce_mean_group'.}
}
\description{
Print method for CCE-MG estimator
}
