% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_ardl_mg.R
\name{print.cs_ardl_mg}
\alias{print.cs_ardl_mg}
\title{Print method for cs_ardl_mg objects}
\usage{
\method{print}{cs_ardl_mg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A \code{cs_ardl_mg} object.}

\item{digits}{Number of digits to print.}

\item{...}{Not used.}
}
\value{
\code{invisible(x)}.
}
\description{
Compact display of key results.
}
