% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_ardl_mg.R
\name{print.summary.cs_ardl_mg}
\alias{print.summary.cs_ardl_mg}
\title{Print method for summary.cs_ardl_mg}
\usage{
\method{print}{summary.cs_ardl_mg}(x, digits = max(3L, getOption("digits") - 3L), signif.stars = TRUE, ...)
}
\arguments{
\item{x}{A \code{summary.cs_ardl_mg} object.}

\item{digits}{Number of digits.}

\item{signif.stars}{Logical; print significance stars.}

\item{...}{Not used.}
}
\value{
\code{invisible(x)}.
}
\description{
Nicely formats the summary output.
}
