% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_methods.R
\name{print.summary.csdm_fit}
\alias{print.summary.csdm_fit}
\title{Print summary of csdm_fit model, including CD and CD* statistics if available}
\usage{
\method{print}{summary.csdm_fit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A summary.csdm_fit object.}

\item{digits}{Number of digits to print.}

\item{...}{Further arguments passed to methods.}
}
\description{
Print summary of csdm_fit model, including CD and CD* statistics if available
}
\details{
If both classic and bias-corrected CD statistics are available, both will be printed side-by-side in the summary output. See [cd_test()] for details.
}
