% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce_mean_group.R
\name{summary.cce_mean_group}
\alias{summary.cce_mean_group}
\title{Summary method for CCE-MG estimator}
\usage{
\method{summary}{cce_mean_group}(object)
}
\arguments{
\item{object}{An object of class 'cce_mean_group'.}
}
\description{
Summary method for CCE-MG estimator
}
