% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_ardl_mg.R
\name{summary.cs_ardl_mg}
\alias{summary.cs_ardl_mg}
\title{Summary for cs_ardl_mg objects}
\usage{
\method{summary}{cs_ardl_mg}(object, z_dist = TRUE, ...)
}
\arguments{
\item{object}{A \code{cs_ardl_mg} object.}

\item{z_dist}{Logical; if \code{TRUE}, use standard normal for p-values (default).
If \code{FALSE}, uses a t-approximation with df = N - 1.}

\item{...}{Not used.}
}
\value{
An object of class \code{summary.cs_ardl_mg}.
}
\description{
Adds test statistics and p-values to MG coefficients.
}
