% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_extract.R
\name{update.csdm_fit}
\alias{update.csdm_fit}
\title{Update a fitted panel model}
\usage{
\method{update}{csdm_fit}(object, formula., ..., evaluate = TRUE)
}
\arguments{
\item{object}{A csdm_fit object.}

\item{formula.}{Formula update.}

\item{...}{Named arguments replacing the original call arguments.}

\item{evaluate}{Evaluate the updated call.}
}
\description{
Update a fitted panel model
}
