% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_ardl_mg.R
\name{vcov.cs_ardl_mg}
\alias{vcov.cs_ardl_mg}
\title{Variance-covariance for cs_ardl_mg}
\usage{
\method{vcov}{cs_ardl_mg}(object, type = c("long_run", "short_run"), ...)
}
\arguments{
\item{object}{A \code{cs_ardl_mg} object.}

\item{type}{One of \code{"long_run"} (default) or \code{"short_run"}.}

\item{...}{Not used.}
}
\value{
A variance-covariance matrix with appropriate row/column names.
}
\description{
Diagonal VCOV based on MG standard errors.
}
