% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm.R
\name{csdm}
\alias{csdm}
\title{Unified front door for csdm estimators}
\usage{
csdm(
  formula,
  data,
  id,
  time,
  model = c("mg", "cce", "dcce", "cs_ardl", "cs_ecm", "cs_dl"),
  csa = csdm_csa(),
  lr = csdm_lr(),
  pooled = csdm_pooled(),
  trend = c("none", "unit", "pooled"),
  fullsample = FALSE,
  mgmissing = FALSE,
  vcov = csdm_vcov(),
  ...
)
}
\description{
Unified front door for csdm estimators
}
