% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm.R
\name{csdm}
\alias{csdm}
\title{Bayesian catch saturation-depletion model}
\usage{
csdm(
  catch = catch,
  effort = effort,
  biomass = NULL,
  q = ~1,
  s = ~1,
  data,
  lower = c(B0 = 100, q = 1e-10, s = 1),
  upper = c(B0 = 1e+10, q = 0.001, s = 100),
  priors = NULL,
  ...
)
}
\arguments{
\item{catch}{\strong{[name]} Variable name for catch (catch).}

\item{effort}{\strong{[name]} Variable name for effort.}

\item{biomass}{\strong{[name]} Formula for biomass. If not provided, inferred
from the formula for parameter q.}

\item{q}{\strong{[formula]} Formula for parameter q.}

\item{s}{\strong{[formula]} Formula for parameter s.}

\item{data}{\strong{[data.frame]} The timeseries data to fit the model with.}

\item{lower}{\strong{[named numeric vector]} Lower bounds of catch estimate.}

\item{upper}{\strong{[named numeric vector]} Upper bounds of catch estimate.}

\item{priors}{\strong{[NA]} List of priors.}

\item{...}{Additional arguments to pass to \link[rstan:stanmodel-method-sampling]{sampling}.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Catch saturation-depletion model.
}
\details{
NA
}
\examples{
\dontrun{
delury_model <- csdm(catch = cpue,
                     effort = traps,
                     biomass = ~1,
                     q = ~1,
                     s = ~1,
                     data = delury,
                     prior = list(B0_logmean_prior_sd = log(1000),
                                  B0_group_prior_logsd = log(100)),
                     upper =  c(B0 = 1e10, q = 1e-3, s = 100),
                     chains = 4,
                     iter = 1000)
}

}
