% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm.R
\name{csdm}
\alias{csdm}
\title{Bayesian catch saturation-depletion model}
\usage{
csdm(x, y, ...)
}
\arguments{
\item{x}{To change}

\item{y}{To change}

\item{...}{To change}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Catch saturation-depletion model.
}
