% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_specs.R
\name{csdm_lr}
\alias{csdm_lr}
\title{Specification: Long-run configuration}
\usage{
csdm_lr(
  vars = NULL,
  type = c("none", "ecm", "ardl", "csdl"),
  ylags = 0,
  xdlags = 0,
  options = list()
)
}
\arguments{
\item{vars}{Reserved for future use.}

\item{type}{One of c("none","ecm","ardl","csdl").}

\item{ylags}{Integer >= 0. Within-unit lags of the dependent variable to include
when supported by the chosen model/type.}

\item{xdlags}{Integer >= 0. Scalar distributed lags to apply to each RHS regressor
when supported by the chosen model/type.}

\item{options}{Reserved for future use.}
}
\value{
A spec object (list) used by csdm().
}
\description{
Specification: Long-run configuration
}
