% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_specs.R
\name{csdm_vcov}
\alias{csdm_vcov}
\title{Specification: Variance-covariance for MG output (stub)}
\usage{
csdm_vcov(type = c("mg", "np", "nw", "wpn", "ols"), ...)
}
\arguments{
\item{type}{One of c("mg","np","nw","wpn","ols").}

\item{...}{Reserved for future use.}
}
\value{
A spec object (list) used by csdm().
}
\description{
Specification: Variance-covariance for MG output (stub)
}
