% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_methods.R
\name{summary.csdm_fit}
\alias{summary.csdm_fit}
\title{Summarize csdm_fit model results, including CD and CD* statistics}
\usage{
\method{summary}{csdm_fit}(object, digits = 4, ...)
}
\arguments{
\item{object}{A csdm_fit model object.}

\item{digits}{Number of digits to print.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of class 'summary.csdm_fit'.
}
\description{
Summarize csdm_fit model results, including CD and CD* statistics
}
\details{
If the model was fit with both classic and bias-corrected CD statistics available, both will be reported in the summary output. See [cd_test()] for details on the CD-type selection and methodology.
}
