% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_methods.R
\name{summary.csdm_fit}
\alias{summary.csdm_fit}
\title{Summarize csdm_fit model results}
\usage{
\method{summary}{csdm_fit}(object, digits = 4, ...)
}
\arguments{
\item{object}{A csdm_fit model object.}

\item{digits}{Number of digits to print.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of class 'summary.csdm_fit'.
}
\description{
Summarize csdm_fit model results
}
\details{
The summary displays classic Pesaran CD test statistics. For additional CD diagnostics
(CDw, CDw+, CD*), use `cd_test()` on the fitted model object.
}
