% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdm_methods.R
\name{vcov.csdm_fit}
\alias{vcov.csdm_fit}
\title{Extract coefficient covariance matrix from a fitted csdm model}
\usage{
\method{vcov}{csdm_fit}(object, ...)
}
\arguments{
\item{object}{A fitted object of class \code{csdm_fit}.}

\item{...}{Currently unused.}
}
\value{
A numeric variance-covariance matrix aligned with \code{coef(object)}
for models where this is available.
}
\description{
Extract coefficient covariance matrix from a fitted csdm model
}
\seealso{
\code{\link[=coef.csdm_fit]{coef.csdm_fit()}}, \code{\link[=summary.csdm_fit]{summary.csdm_fit()}}
}
